LF-4124 get rid of ref in other components
